+++ src/components/user-info/presentation.tsx
@@ -8,9 +8,9 @@ interface Props {
   bio: string;
 }
 
-export const MainInfo = ({ firstname, lastName, imageUrl, headliner, bio }: Props) => {
+export const Presentation = ({ firstname, lastName, imageUrl, headliner, bio }: Props) => {
   return (
-    <section className="w-full text-light-blue dark:text-white rounded-lg flex flex-col items-center gap-4 md:flex-row py-7 px-3">
+    <section className="w-full text-dark-blue dark:text-white rounded-lg flex flex-col items-center gap-4 md:flex-row py-7 px-3">
       <Image
         src={imageUrl ?? ''}
         className="rounded-full w-[200px] h-[200px] border-2 border-light-blue dark:border-white"
@@ -21,10 +21,10 @@ export const MainInfo = ({ firstname, lastName, imageUrl, headliner, bio }: Prop
       <div>
         <h1 className="text-xl md:text-2xl font-bold">
           Hi there 👋! My name is{' '}
-          <span className="text-light-green">
+          <span className="text-blue-500 dark:text-light-green font-mono">
             {firstname} {lastName}
           </span>{' '}
-          I&apos;m <span className="text-light-green">{headliner}</span>
+          I&apos;m <span className="text-blue-500 dark:text-light-green">{headliner}</span>
         </h1>
         <h4 className="md:text-lg font-bold my-6">A brief introduction about me...</h4>
         <p className="md:text-lg text-justify">{bio}</p>
