+++ src/components/user-info/social-media.tsx
@@ -1,4 +1,3 @@
-import clsx from 'clsx';
 import {
   FaFacebook,
   FaGithub,
@@ -9,6 +8,7 @@ import {
 } from 'react-icons/fa6';
 
 import { SocialMedia } from '@/interfaces/profile';
+import { cn } from '@/lib/utils';
 
 interface Props {
   socialMedia: SocialMedia;
@@ -38,7 +38,7 @@ export const SocialMediaInfo = ({ socialMedia, className }: Props) => {
   };
 
   return (
-    <div className={clsx('flex gap-4 flex-wrap px-5', className)}>
+    <div className={cn('flex gap-4 flex-wrap px-5', className)}>
       {validSocialMedia.map((sm) => (
         <a
           key={sm}
@@ -47,7 +47,7 @@ export const SocialMediaInfo = ({ socialMedia, className }: Props) => {
           rel="noopener noreferrer"
           className="flex flex-col gap-2 items-center hover:underline hover:underline-offset-4"
         >
-          {getIcon(sm, 'w-6 h-6 text-light-green dark:text-light-gold')}
+          {getIcon(sm, 'w-6 h-6 text-blue-500 dark:text-light-gold')}
           <span>{sm.at(0)!.toUpperCase() + sm.slice(1)}</span>
         </a>
       ))}
